+++ services/ui-src/src/measures/2023/shared/globalValidations/index.ts
@@ -15,7 +15,7 @@ export * from "./validateEqualCategoryDenominators";
 export * from "shared/globalValidations/validateEqualQualifierDenominators";
 export * from "./validateFfsRadioButtonCompletion";
 export * from "./validateRateNotZero";
-export * from "./validateRateZero";
+export * from "shared/globalValidations/validateRateZero";
 export * from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 export * from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
 export * from "shared/globalValidations/validateOneQualDenomHigherThanOtherDenomOMS";
