+++ services/ui-src/src/measures/2023/shared/util/validationsMock.tsx
@@ -10,7 +10,7 @@ import * as validateEqualCategoryDenominators from "measures/2023/shared/globalV
 import * as validateEqualQualifierDenominators from "shared/globalValidations/validateEqualQualifierDenominators";
 import * as validateFfsRadioButtonCompletion from "measures/2023/shared/globalValidations/validateFfsRadioButtonCompletion";
 import * as validateRateNotZero from "measures/2023/shared/globalValidations/validateRateNotZero";
-import * as validateRateZero from "measures/2023/shared/globalValidations/validateRateZero";
+import * as validateRateZero from "shared/globalValidations/validateRateZero";
 import * as validateNumeratorsLessThanDenominators from "shared/globalValidations/validateNumeratorsLessThanDenominators";
 import * as validateOneCatRateHigherThanOtherCat from "shared/globalValidations/validateOneCatRateHigherThanOtherCat";
 import * as validateOneQualDenomHigherThanOtherDenomOMS from "shared/globalValidations/validateOneQualDenomHigherThanOtherDenomOMS";
